Add environment for integration tests
This adds a new env argument for integration tests which is triggered by running yarn test-app --env=integration. This is not necessarily a "full" environment, but is run inside the node env, and also filters which tests get run. The reason we want a new environment is so that we can control when and how we spin these tests up, for example we may want some other runner which wraps and runs just these integration tests. It also allows us to easily split off integration tests into another environment.