Pin enzyme-to-json@3.3.1
Version 3.3.2 was published to the npm registry briefly before being unpublished. Unfortunately some references remain in registries, and running yarn add enzyme-to-json@^3.3.1 will cause yarn to error. We can temporarily pin the version to get around this, and I've also filed https://github.com/adriantoine/enzyme-to-json/issues/100 in the meantime.